+++ vimms/Controller/__init__.py
@@ -5,4 +5,4 @@ from vimms.Controller.misc import *
 from vimms.Controller.roi import *
 from vimms.Controller.topN import *
 from vimms.Controller.targeted import *
-from vimms.Controller.shell import *
+from vimms.Controller.abc import *
